18/19 patch 2
this is a very important patch,
in this commit we completely reworked how the app works. you are now longer loading the menu screen and popping back to the actual activity but instead the actual activity (main_vp) instantely loads up and only pops you to the menu if this is your first start and the class isn't set yet.
// this requires the settings tab to work in the future. just letting you know im looking into this next. //
nevertheless this was needed to continuously increase the performance and loading speed of the app. it now loads up to 300% faster and swaps days with a speed of less than a half second on my S9. also the loading animation now got fully implemented but sometimes isn't even showing due to the loading speed of the app. might show more frequently on slower devices (untested). futhermore the ugly white screen when starting the app got removed, this makes the overall UX much better.

TODO:
- design optimizations: thinking of a complete rework, but not sure yet.
- message of the day crawler: thinking of a way to implement that nicely.